Enable schema validation and fix type mismatches

+++ src/main/resources/db/migration/V1_1__new_application2.sql
@@ -1,6 +1,6 @@
 CREATE TABLE door_topic(
-	topics_id INT,
-	door_id INT,
+	topics_id BIGINT,
+	door_id BIGINT,
 	FOREIGN KEY (topics_id) REFERENCES topic(id),
 	FOREIGN KEY (door_id) REFERENCES door(id)
 )
